+++ packages/preprocessor/CHANGELOG.md
@@ -1,5 +1,7 @@
 ### HEAD
 
+* Add `postcss-apply`
+* Add `postcss-color-function`
 * Add `debug` option to pass an `postcss-debug` instance
 * Allow usage of `transform` and `onImport` in `postcss-easy-import`
 * Remove `beforeLint` feature